feat #FP-I26: moved course related apis calls to course services
