+++ src/course_enrollment_script/services/courseService.ts
@@ -271,3 +271,91 @@ export async function publishContent(identifier: string): Promise<void> {
         throw error;
     }
 }
+
+export async function searchLearnerProfile(profileCode: string): Promise<string | null> {
+    const headers = {
+        'Content-Type': 'application/json',
+        'X-Channel-Id': config.channelId,
+        'Authorization': config.apiAuthKey,
+        'x-authenticated-user-token': globalConfig.creatorUserToken
+    };
+
+    const body = {
+        request: {
+            filters: {
+                code: profileCode
+            }
+        }
+    };
+
+    try {
+        const response = await axios.post(`${config.baseUrl}${routes.searchCourse}`, body, { headers });
+        const content = response.data.result.content;
+        const count = response.data.result.count;
+
+        if (count === 0 || !content || content.length === 0) {
+            console.log(`No content found for learner profile ${profileCode}`);
+            return null;
+        }
+
+        const profile = content[0];
+        if (profile.contentType !== 'Resource' || !profile.children) {
+            console.log(`Invalid learner profile ${profileCode}: wrong content type or missing required fields`);
+            return null;
+        }
+
+        return profile.identifier;
+    } catch (error) {
+        console.error(`Error searching for learner profile ${profileCode}:`, error);
+        return null;
+    }
+}
+
+export async function getProfileCourses(profileId: string): Promise<string[]> {
+    const headers = {
+        'Content-Type': 'application/json',
+        'X-Channel-Id': config.channelId,
+        'Authorization': config.apiAuthKey,
+        'x-authenticated-user-token': globalConfig.creatorUserToken
+    };
+
+    try {
+        const response = await axios.get(`${config.baseUrl}${routes.readContent}/${profileId}`, { headers });
+        const children = response.data.result.content.children;
+
+        if (!children || !Array.isArray(children)) {
+            console.log(`No courses found in learner profile ${profileId}`);
+            return [];
+        }
+
+        return children.map(child => child.identifier);
+    } catch (error) {
+        console.error(`Error getting courses for profile ${profileId}:`, error);
+        return [];
+    }
+}
+
+export async function getCourseNodeIds(courseIds: string[]): Promise<{ [nodeId: string]: string }> {
+    const nodeIdToCourseCodeMap: { [nodeId: string]: string } = {};
+    for (const courseId of courseIds) {
+        const headers = {
+            'Content-Type': 'application/json',
+            'X-Channel-Id': config.channelId,
+            'Authorization': config.apiAuthKey,
+            'x-authenticated-user-token': globalConfig.creatorUserToken
+        };
+
+        try {
+            const response = await axios.get(`${config.baseUrl}${routes.readContent}/${courseId}`, { headers });
+            const {status, contentType, code} = response.data.result.content;
+            if(status !== 'Live' || contentType !== 'Course' ||!code) {
+                console.log(`Invalid course ${courseId}: wrong status or content type or missing required fields`);
+                continue;
+            }
+            nodeIdToCourseCodeMap[courseId] = code;
+        } catch (error) {
+            console.error(`Error searching for course ${courseId}:`, error);
+        }
+    }
+    return nodeIdToCourseCodeMap;
+}
